new:Add backend route event-api.js, organized files for sales, fixed bugs in create sales
