+++ backend/routes/events-api.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', (req, res) => {
+  res.json({ message: 'Events endpoint' });
+});
+
+module.exports = router;
