+++ frontend/src/components/EditSalesDialog.jsx
@@ -110,7 +110,7 @@ export function EditSalesDialog({ open, onOpenChange, defaultValues, fetchSales 
     console.log(payload)
 
     try {
-      const res = await fetch("http://localhost:8080/api/users/updateSale", {
+      const res = await fetch("http://localhost:8080/api/sales/updateSale", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
         body: JSON.stringify(payload),
